Move source to src
