+++ package.json
@@ -6,7 +6,7 @@
   "scripts": {
     "test": "mocha -r esm",
     "test:watch": "nodemon --watch . --exec 'npm test || true'",
-    "build:interface": "rollup index.js -f es --file build/interface.js",
+    "build:interface": "rollup src/index.js -f es --file build/interface.js",
     "build": "npm run build:interface && ./build.sh less_basic_model_for_js.R build/interface.js build/squire.js",
     "e2e": "npm run build && node -r esm e2e/test.js"
   },
